refactor import

+++ src/main/java/com/example/springcrudpharse02notecollector/controller/UserController.java
@@ -1,9 +1,10 @@
 package com.example.springcrudpharse02notecollector.controller;
 
 import com.example.springcrudpharse02notecollector.customStatusCode.selectedUserErrorStatus;
-import com.example.springcrudpharse02notecollector.dto.UserDTO;
+import com.example.springcrudpharse02notecollector.dto.impl.UserDTO;
 import com.example.springcrudpharse02notecollector.dto.UserStatus;
 import com.example.springcrudpharse02notecollector.exception.DataPersistException;
+import com.example.springcrudpharse02notecollector.exception.UserNotFoundException;
 import com.example.springcrudpharse02notecollector.service.UserService;
 import com.example.springcrudpharse02notecollector.util.AppUtil;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,6 +16,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.regex.Pattern;
 
 @RestController
 @RequestMapping("api/v1/user")
@@ -32,10 +34,10 @@ public class UserController {
             @RequestPart("password") String password,
             @RequestPart("profilePicture") MultipartFile profilePicture) {
         //Profile picture -----> Base64
-        String base64ProPic = "";
+//        String base64ProPic = "";
         try {
             byte[] bytesProPic = profilePicture.getBytes();
-            base64ProPic = AppUtil.profilePicBase64(bytesProPic);
+            String base64ProPic = AppUtil.profilePicBase64(bytesProPic);
 
             //UserId generate
             var userId = AppUtil.generateUserId();
@@ -60,12 +62,14 @@ public class UserController {
 
     @GetMapping(value = "/{userId}", produces = MediaType.APPLICATION_JSON_VALUE)
     public UserStatus getSelectedUser(@PathVariable("userId") String userId) {
-        if (userId.isEmpty() || userId == null){
+        String regexUserId = "^USER-[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$";
+        Pattern regexPattern = Pattern.compile(regexUserId);
+        var matcherUserId = regexPattern.matcher(userId);
+        if (!matcherUserId.matches()){
             //meke response eka thawa application ekakata ynawa kiyla hithamu.e nisa apita eka http status code walin handle krnna bh.e nisa apita pluwn custom status code ekk hadala eka return kranna.
             return new selectedUserErrorStatus(1,"User ID not valid!");
         }
         return userService.getUser(userId);
-
     }
 
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
@@ -73,12 +77,25 @@ public class UserController {
         return userService.getAllUsers();
     }
 
-    @ResponseStatus(HttpStatus.NO_CONTENT)
+//    @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping(value = "/{userId}")
-    public void deleteUser(@PathVariable("userId") String userId) {
-        userService.deleteUser(userId);
+    public ResponseEntity<Void> deleteUser(@PathVariable("userId") String userId) {
+        String regexUserId = "^USER-[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$";
+        Pattern regexPattern = Pattern.compile(regexUserId);
+        var matcherUserId = regexPattern.matcher(userId);
+        try {
+            if (!matcherUserId.matches()){
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+            userService.deleteUser(userId);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }catch (UserNotFoundException e){
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
-
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PutMapping(value = "/{userId}")
     public void updateUser(@PathVariable("userId") String userId, @RequestPart("firstName") String firstName, @RequestPart("lastName") String lastName, @RequestPart("email") String email, @RequestPart("password") String password, @RequestPart("profilePicture") MultipartFile profilePicture) throws IOException {
